refactor: move player profile query slice into core application
