+++ backend/src/core/application/use_cases/__init__.py
@@ -1,5 +1,6 @@
 from .get_nationalities_usecase import GetNationalitiesUseCase
 from .get_penas_usecase import GetPenasUseCase
+from .get_player_profile_usecase import GetPlayerProfileUseCase
 from .get_season_match_insights_usecase import GetSeasonMatchInsightsUseCase
 from .season_match_insights_errors import (
     InvalidSeasonInsightsDataError,
@@ -10,6 +11,7 @@ from .season_match_insights_errors import (
 __all__ = [
     "GetNationalitiesUseCase",
     "GetPenasUseCase",
+    "GetPlayerProfileUseCase",
     "GetSeasonMatchInsightsUseCase",
     "InvalidSeasonInsightsDataError",
     "PenaSeasonNotFoundError",
