+++ backend/tests/test_players_controller.py
@@ -3,8 +3,8 @@ from api.interface.controller.v1 import players_controller
 from api.interface.controller.v1.model.request.players_request import PlayerUpdateRequest
 from auth.dependencies import require_user
 from auth.session import SessionData
+from core.application.models import PenaInfo, PlayerProfile
 from fastapi import HTTPException
-from persistence.application.use_cases.get_player_profile_usecase import PenaInfo, PlayerProfile
 from persistence.application.use_cases.update_player_profile_usecase import (
     InvalidNationalityError as PlayerInvalidNationalityError,
 )
